introduce de-metas-common-manufacturing project; refactor
