+++ backend/de.metas.manufacturing/src/main/java/de/metas/manufacturing/order/exportaudit/ExportTransactionId.java
@@ -40,7 +40,7 @@ public final class ExportTransactionId
 	}
 
 	@JsonCreator
-	public static ExportTransactionId ofString(final String value)
+	public static ExportTransactionId ofString(@NonNull final String value)
 	{
 		return new ExportTransactionId(value);
 	}
