+++ misc/de-metas-common/de-metas-common-manufacturing/src/main/java/de/metas/common/manufacturing/JsonRequestHULookup.java
@@ -1,16 +1,12 @@
-package de.metas.manufacturing.rest_api;
+package de.metas.common.manufacturing;
 
-import java.time.ZonedDateTime;
+import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 
-import de.metas.common.rest_api.JsonQuantity;
-import de.metas.material.planning.pporder.PPOrderId;
-import de.metas.util.time.SystemTime;
 import lombok.Builder;
-import lombok.Builder.Default;
 import lombok.NonNull;
 import lombok.Value;
 
@@ -39,21 +35,17 @@ import lombok.Value;
 @Value
 @Builder
 @JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY, getterVisibility = JsonAutoDetect.Visibility.NONE, isGetterVisibility = JsonAutoDetect.Visibility.NONE, setterVisibility = JsonAutoDetect.Visibility.NONE)
-@JsonDeserialize(builder = JsonRequestReceiveFromManufacturingOrder.JsonRequestReceiveFromManufacturingOrderBuilder.class)
-public class JsonRequestReceiveFromManufacturingOrder
+@JsonDeserialize(builder = JsonRequestHULookup.JsonRequestHULookupBuilder.class)
+public class JsonRequestHULookup
 {
 	@NonNull
-	PPOrderId orderId;
+	String lotNumber;
 
 	@NonNull
-	JsonQuantity qtyToReceive;
-
-	@NonNull
-	@Default
-	ZonedDateTime date = SystemTime.asZonedDateTime();
+	LocalDate bestBeforeDate;
 
 	@JsonPOJOBuilder(withPrefix = "")
-	public static class JsonRequestReceiveFromManufacturingOrderBuilder
+	public static class JsonRequestHULookupBuilder
 	{
 	}
 }
