+++ misc/de-metas-common/de-metas-common-manufacturing/src/main/java/de/metas/common/manufacturing/JsonRequestManufacturingOrdersReport.java
@@ -1,15 +1,14 @@
-package de.metas.manufacturing.rest_api;
+package de.metas.common.manufacturing;
 
-import org.eevolution.api.BOMComponentType;
+import java.util.List;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
-import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 
-import de.metas.common.rest_api.JsonQuantity;
-import de.metas.common.shipping.JsonProduct;
 import lombok.Builder;
+import lombok.NonNull;
+import lombok.Singular;
 import lombok.Value;
 
 /*
@@ -36,16 +35,20 @@ import lombok.Value;
 
 @Value
 @Builder
-@JsonAutoDetect(fieldVisibility = Visibility.ANY, getterVisibility = Visibility.NONE, isGetterVisibility = Visibility.NONE, setterVisibility = Visibility.NONE)
-@JsonDeserialize(builder = JsonResponseManufacturingOrderBOMLine.JsonResponseManufacturingOrderBOMLineBuilder.class)
-public class JsonResponseManufacturingOrderBOMLine
+@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY, getterVisibility = JsonAutoDetect.Visibility.NONE, isGetterVisibility = JsonAutoDetect.Visibility.NONE, setterVisibility = JsonAutoDetect.Visibility.NONE)
+@JsonDeserialize(builder = JsonRequestManufacturingOrdersReport.JsonRequestManufacturingOrdersReportBuilder.class)
+public class JsonRequestManufacturingOrdersReport
 {
-	BOMComponentType componentType;
-	JsonProduct product;
-	JsonQuantity qty;
+	@NonNull
+	@Singular
+	List<JsonRequestReceiveFromManufacturingOrder> receipts;
+
+	@NonNull
+	@Singular
+	List<JsonRequestIssueToManufacturingOrder> issues;
 
 	@JsonPOJOBuilder(withPrefix = "")
-	public static class JsonResponseManufacturingOrderBOMLineBuilder
+	public static class JsonRequestManufacturingOrdersReportBuilder
 	{
 	}
 }
