+++ misc/de-metas-common/de-metas-common-manufacturing/src/main/java/de/metas/common/manufacturing/JsonResponseManufacturingOrder.java
@@ -1,13 +1,21 @@
-package de.metas.manufacturing.rest_api;
+package de.metas.common.manufacturing;
 
-import java.time.LocalDate;
+import java.time.ZonedDateTime;
+import java.util.List;
+
+import javax.annotation.Nullable;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
+import com.fasterxml.jackson.annotation.JsonAutoDetect.Visibility;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonPOJOBuilder;
 
+import de.metas.common.rest_api.JsonMetasfreshId;
+import de.metas.common.rest_api.JsonQuantity;
+import de.metas.common.shipping.JsonProduct;
 import lombok.Builder;
 import lombok.NonNull;
+import lombok.Singular;
 import lombok.Value;
 
 /*
@@ -34,18 +42,35 @@ import lombok.Value;
 
 @Value
 @Builder
-@JsonAutoDetect(fieldVisibility = JsonAutoDetect.Visibility.ANY, getterVisibility = JsonAutoDetect.Visibility.NONE, isGetterVisibility = JsonAutoDetect.Visibility.NONE, setterVisibility = JsonAutoDetect.Visibility.NONE)
-@JsonDeserialize(builder = JsonRequestHULookup.JsonRequestHULookupBuilder.class)
-public class JsonRequestHULookup
+@JsonAutoDetect(fieldVisibility = Visibility.ANY, getterVisibility = Visibility.NONE, isGetterVisibility = Visibility.NONE, setterVisibility = Visibility.NONE)
+@JsonDeserialize(builder = JsonResponseManufacturingOrder.JsonResponseManufacturingOrderBuilder.class)
+public class JsonResponseManufacturingOrder
 {
 	@NonNull
-	String lotNumber;
+	JsonMetasfreshId orderId;
 
 	@NonNull
-	LocalDate bestBeforeDate;
+	String documentNo;
+
+	@Nullable
+	String description;
+
+	@NonNull
+	ZonedDateTime dateOrdered;
+	@NonNull
+	ZonedDateTime dateStartSchedule;
+
+	@NonNull
+	JsonProduct finishGoodProduct;
+	@NonNull
+	JsonQuantity qtyToProduce;
+
+	@NonNull
+	@Singular
+	List<JsonResponseManufacturingOrderBOMLine> components;
 
 	@JsonPOJOBuilder(withPrefix = "")
-	public static class JsonRequestHULookupBuilder
+	public static class JsonResponseManufacturingOrderBuilder
 	{
 	}
 }
